unify formatters for nano s and nano x
