+++ src/stellar_types.h
@@ -82,19 +82,21 @@
 #define NETWORK_TYPE_TEST    1
 #define NETWORK_TYPE_UNKNOWN 2
 
-#define XDR_OPERATION_TYPE_CREATE_ACCOUNT       0
-#define XDR_OPERATION_TYPE_PAYMENT              1
-#define XDR_OPERATION_TYPE_PATH_PAYMENT         2
-#define XDR_OPERATION_TYPE_MANAGE_SELL_OFFER    3
-#define XDR_OPERATION_TYPE_CREATE_PASSIVE_OFFER 4
-#define XDR_OPERATION_TYPE_SET_OPTIONS          5
-#define XDR_OPERATION_TYPE_CHANGE_TRUST         6
-#define XDR_OPERATION_TYPE_ALLOW_TRUST          7
-#define XDR_OPERATION_TYPE_ACCOUNT_MERGE        8
-#define XDR_OPERATION_TYPE_INFLATION            9
-#define XDR_OPERATION_TYPE_MANAGE_DATA          10
-#define XDR_OPERATION_TYPE_BUMP_SEQUENCE        11
-#define XDR_OPERATION_TYPE_MANAGE_BUY_OFFER     12
+typedef enum {
+    XDR_OPERATION_TYPE_CREATE_ACCOUNT = 0,
+    XDR_OPERATION_TYPE_PAYMENT = 1,
+    XDR_OPERATION_TYPE_PATH_PAYMENT = 2,
+    XDR_OPERATION_TYPE_MANAGE_SELL_OFFER = 3,
+    XDR_OPERATION_TYPE_CREATE_PASSIVE_OFFER = 4,
+    XDR_OPERATION_TYPE_SET_OPTIONS = 5,
+    XDR_OPERATION_TYPE_CHANGE_TRUST = 6,
+    XDR_OPERATION_TYPE_ALLOW_TRUST = 7,
+    XDR_OPERATION_TYPE_ACCOUNT_MERGE = 8,
+    XDR_OPERATION_TYPE_INFLATION = 9,
+    XDR_OPERATION_TYPE_MANAGE_DATA = 10,
+    XDR_OPERATION_TYPE_BUMP_SEQUENCE = 11,
+    XDR_OPERATION_TYPE_MANAGE_BUY_OFFER = 12,
+} xdr_operation_type_e;
 
 #define SIGNER_KEY_TYPE_ED25519     0
 #define SIGNER_KEY_TYPE_PRE_AUTH_TX 1
